Fix warnings and build oddities

+++ adpcm/adpcm-lib.c
@@ -1005,7 +1005,7 @@ int adpcm_decode_block (int16_t *outbuf, const uint8_t *inbuf, size_t inbufsize,
 
 int adpcm_decode_block_ex (int16_t *outbuf, const uint8_t *inbuf, size_t inbufsize, int channels, int bps)
 {
-    int samples = 1, ch;
+    size_t samples = 1, ch;
     int32_t pcmdata[2];
     int8_t index[2];
